+++ survey/models.py
@@ -42,6 +42,10 @@ class Question(TimeStampedModel):
         help_text='The order in which the question appears in the survey.',
     )
 
+    @property
+    def choices_with_additional_routing(self):
+        return [c for c in self.choices.all() if c.additional_routing]
+
     def __str__(self):
         return self.title
 
@@ -55,6 +59,7 @@ class Choice(TimeStampedModel):
         (NO_ROUTING, 'None, go to next question'),
         (JUMP, 'Jump to a different question'),
     )
+    ADDITIONAL_ROUTING_ERROR = 'Multi-select questions cannot have more than one choice with additional routing.'
 
     question = models.ForeignKey(Question, related_name='choices', on_delete=models.CASCADE)
     label = models.CharField(
@@ -92,7 +97,7 @@ class Choice(TimeStampedModel):
     )
 
     def clean_fields(self, *args, **kwargs):
+        super().clean_fields(*args, **kwargs)
         if self.additional_routing == self.JUMP:
             if not self.question_to_jump_to:
                 raise ValidationError({'question_to_jump_to': ['This field is required.']})
-        return super().clean_fields(*args, **kwargs)