+++ CHANGELOG.md
@@ -2,14 +2,25 @@
 
 ## Pre release
 ### Bugs fixed
-- KLS-348 - View attributes renamed to match filtersets
 
 ### Enhancements
 - KLS-396 - Patch django to 3.2.18
-- KLS-265 - Patch certify to 2022.12.17
+- KLS-385 - Add validation for additional routing on multi-select survey questions
+
+## [2.10.0](https://github.com/uktrade/directory-api/releases/tag/2.10.0)
+[Full Changelog](https://github.com/uktrade/directory-api/compare/2.9.0...2.10.0)
+### Bugs fixed
+- KLS-348 - View attributes renamed to match filtersets
+
+### Enhancements
 - KLS-339 - Make label and value fields on Choice model not unique
 - KLS-290 - Add additional_routing logic to survey
 - Update black, blacken-docs and isort versions to latest
+
+## [2.9.0](https://github.com/uktrade/directory-api/releases/tag/2.9.0)
+[Full Changelog](https://github.com/uktrade/directory-api/compare/2.8.0...2.9.0)
+### Enhancements
+- KLS-265 - Patch certify to 2022.12.17
 
 ## [2.8.0](https://github.com/uktrade/directory-api/releases/tag/2.8.0)
 [Full Changelog](https://github.com/uktrade/directory-api/compare/2.7.0...2.8.0)